Enhance ticket list filters, comment deletion, and UX
- Added advanced filtering and search to the ticket list (by query, priority, status, including combined status filters).
- Updated ticket list to use a filteredTickets collection for dynamic filtering.
- Improved ticket assignment logic: prevent self-assignment, show toasts for assignment errors.
- Added toast notifications to ticket details for user feedback.
- Enabled comment deletion for authors/admins; added UI for delete action.
- Refactored repository interfaces to use int IDs instead of long.
- Implemented DeleteCommentAsync in the ticket service.
- Updated localization files with new strings for filters, assignment, and errors.

+++ ServiceDeskSystem/Data/Repository/CommentRepository.cs
@@ -13,14 +13,14 @@ namespace ServiceDeskSystem.Data.Repository
 
         protected override DbSet<Comment> DbSet => this.Context.Comments;
 
-        public Comment? GetByIdWithAuthor(long id)
+        public Comment? GetByIdWithAuthor(int id)
         {
             return this.Context.Comments
                 .Include(c => c.Author)
                 .FirstOrDefault(c => c.Id == id);
         }
 
-        public async Task<Comment?> GetByIdWithAuthorAsync(long id)
+        public async Task<Comment?> GetByIdWithAuthorAsync(int id)
         {
             return await this.Context.Comments
                 .Include(c => c.Author)
